added a chance for an image to be found for articles
